refactor: 분산된 date util 함수 date 폴더로 이동 및 적용
