+++ frontend/src/@constants/date.ts
@@ -20,3 +20,6 @@ export const DAY = {
   5: "금요일",
   6: "토요일",
 } as const;
+
+export const WEEKDAYS = ["일", "월", "화", "수", "목", "금", "토"] as const;
+export const MONTHS = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12] as const;
