+++ frontend/src/components/DateDropdown/DateDropdownOption/index.tsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import * as Styled from "@src/components/DateDropdown/DateDropdownMenu/style";
 
 import { DATE } from "@src/@constants/date";
-import { ISOConverter } from "@src/@utils";
+import { ISOConverter } from "@src/@utils/date";
 
 interface Props {
   date: string;
